requirements.txt
flask
flask-cors
keybert
anthropic
python-dotenv
spacy
